# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NavPlat API
  description: Navigation Platform API Documentation - Complete API specification for Postman and other API clients
  version: 1.0.0
  contact:
    name: NavPlat Team
    email: support@navplat.local
servers:
  - url: http://localhost:5000/api
    description: Local Development Gateway
  - url: https://api.navplat.local
    description: Production Gateway
...
paths:
...
  /journeys:
    get:
      tags: [Journeys]
      summary: Get user's journeys
      description: Returns paginated list of journeys for the authenticated user (includes own journeys and shared journeys)
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: startDateFrom
          in: query
          schema:
            type: string
            format: date-time
          description: Filter journeys starting from this date
        - name: startDateTo
          in: query
          schema:
            type: string
            format: date-time
          description: Filter journeys starting until this date
      responses:
        '200':
          description: List of journeys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedJourneys'
        '401':
          description: Unauthorized
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    post:
      tags: [Journeys]
      summary: Create a new journey
      description: Creates a new journey for the authenticated user. Publishes JourneyCreated domain event.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJourneyRequest'
            example:
              startLocation: "Tirana, AL"
              startTime: "2025-04-11T08:15:00Z"
              arrivalLocation: "Durrës, AL"
              arrivalTime: "2025-04-11T09:05:00Z"
              transportType: "Commercial"
              distanceKm: 38.40
      responses:
        '201':
          description: Journey created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateJourneyResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
...
  /journeys/{id}/favorite:
    post:
      tags: [Journeys]
      summary: Add journey to favorites
      description: Adds a journey to user's favorites. Idempotent operation - safe to call multiple times.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Already in favorites (idempotent response)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Journey is already in favorites"
        '204':
          description: Added to favorites
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Journey not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized

    delete:
      tags: [Journeys]
      summary: Remove journey from favorites
      description: Removes a journey from user's favorites
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Removed from favorites
        '404':
          description: Journey not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
...
  /admin/journeys:
    get:
      tags: [Admin]
      summary: Get all journeys with filtering
      description: Returns paginated list of all journeys with advanced filtering options. Requires Admin role.
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: userId
          in: query
          schema:
            type: string
          description: Filter by user ID
        - name: transportType
          in: query
          schema:
            $ref: '#/components/schemas/TransportType'
          description: Filter by transport type
        - name: startDateFrom
          in: query
          schema:
            type: string
            format: date-time
          description: Filter journeys starting from this date
        - name: startDateTo
          in: query
          schema:
            type: string
            format: date-time
          description: Filter journeys starting until this date
        - name: arrivalDateFrom
          in: query
          schema:
            type: string
            format: date-time
          description: Filter journeys arriving from this date
        - name: arrivalDateTo
          in: query
          schema:
            type: string
            format: date-time
          description: Filter journeys arriving until this date
        - name: minDistance
          in: query
          schema:
            type: number
            format: decimal
          description: Minimum distance in kilometers
        - name: maxDistance
          in: query
          schema:
            type: number
            format: decimal
          description: Maximum distance in kilometers
        - name: orderBy
          in: query
          schema:
            type: string
            enum: [UserId, TransportType, DistanceKm, StartTime]
          description: Field to sort by
        - name: direction
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort direction
      responses:
        '200':
          description: List of journeys
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: Total count of journeys matching filters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedJourneys'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin role required
  /admin/statistics:
    get:
      tags: [Admin]
      summary: Get system statistics
      description: Returns overall system statistics (total journeys, users, distance). Requires Admin role.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: System statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Statistics'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin role required
  /admin/statistics/monthly-distance:
    get:
      tags: [Admin]
      summary: Get monthly distance statistics
      description: Returns total distance travelled per user per calendar month. Data from CQRS read model. Requires Admin role.
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: orderBy
          in: query
          schema:
            type: string
            enum: [UserId, TotalDistanceKm]
            default: UserId
          description: Field to sort by
        - name: direction
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort direction
      responses:
        '200':
          description: Monthly distance statistics
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: Total count of records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MonthlyDistance'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin role required
  /admin/users/{userId}/journeys:
    get:
      tags: [Admin]
      summary: Get journeys for a specific user
      description: Returns paginated list of journeys for a specific user. Requires Admin role.
      security:
        - cookieAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: User ID
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: List of journeys for the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedJourneys'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin role required
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: .NavPlat.Auth
      description: HTTP-only cookie containing authentication token
  schemas:
...
    Journey:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        userId:
          type: string
          example: "user-123"
        startLocation:
          type: string
          example: "Tirana, AL"
        startTime:
          type: string
          format: date-time
          example: "2025-04-11T08:15:00Z"
        arrivalLocation:
          type: string
          example: "Durrës, AL"
        arrivalTime:
          type: string
          format: date-time
          example: "2025-04-11T09:05:00Z"
        transportType:
          type: string
          enum: [Commercial, Cargo, Private, Charter, Military, Helicopter, Other]
          example: "Commercial"
        distanceKm:
          type: number
          format: decimal
          example: 38.40
        isFavorite:
          type: boolean
          example: false
        isShared:
          type: boolean
          example: false
        isDailyGoalAchieved:
          type: boolean
          example: false
    CreateJourneyRequest:
      type: object
      required: [startLocation, startTime, arrivalLocation, arrivalTime, transportType, distanceKm]
      properties:
        startLocation:
          type: string
          minLength: 1
          example: "Tirana, AL"
        startTime:
          type: string
          format: date-time
          example: "2025-04-11T08:15:00Z"
        arrivalLocation:
          type: string
          minLength: 1
          example: "Durrës, AL"
        arrivalTime:
          type: string
          format: date-time
          example: "2025-04-11T09:05:00Z"
        transportType:
          type: string
          enum: [Commercial, Cargo, Private, Charter, Military, Helicopter, Other]
          example: "Commercial"
        distanceKm:
          type: number
          format: decimal
          minimum: 0
          example: 38.40
...
    CreateJourneyResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
...
    PagedJourneys:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Journey'
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalCount:
          type: integer
          example: 150
...
    Statistics:
      type: object
      properties:
        totalJourneys:
          type: integer
          example: 1250
        totalUsers:
          type: integer
          example: 45
        totalDistanceKm:
          type: number
          format: decimal
          example: 12500.75
        averageDistanceKm:
          type: number
          format: decimal
          example: 10.00
    MonthlyDistance:
      type: object
      required: [userId, year, month, totalDistanceKm]
      properties:
        userId:
          type: string
          description: User ID
          example: "user-123"
        year:
          type: integer
          description: Calendar year
          example: 2025
        month:
          type: integer
          description: Calendar month (1-12)
          minimum: 1
          maximum: 12
          example: 4
        totalDistanceKm:
          type: number
          format: decimal
          description: Total distance in kilometers for this user in this month
          example: 125.50
...
    TransportType:
      type: string
      enum: [Commercial, Cargo, Private, Charter, Military, Helicopter, Other]
      description: Type of transport used for the journey
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
          example: "https://tools.ietf.org/html/rfc7231#section-6.5.1"
        title:
          type: string
          example: "Bad Request"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "Validation failed"
        instance:
          type: string
          format: uri
          example: "/api/journeys"
        traceId:
          type: string
          example: "0HN1234567890"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Validation errors by field name